+++ changelog.md
@@ -4,6 +4,7 @@
 * Removed Dragon Larva
   * No longer exists in any form
 * Enhanced Blast attack animation and particles
+* Dragon no longer takes knockback
 * Fixed client-server de-sync with dragon anger
 * Fixed dragon being able to be stuck in angered mode forever
 * Lowered chance to strafe when angered and to blast when not